change multiStepForm logic
