+++ frontend/src/interfaces/Tools.ts
@@ -1,0 +1,4 @@
+export interface Tools {
+    name: string;
+    image: File | null;
+}
